+++ packages/experiments/src/components/Text/examples/Text.Ramp.Example.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
-import { Text, VerticalStack } from '@uifabric/experiments';
-import { IFontVariants, IFontSizes, IFontWeights, IFontFamilies, IStyle } from '@uifabric/experiments/lib/Styling';
-import { ISemanticTextColors, IPalette } from '@uifabric/experiments/lib/Styling';
+import { Text, Stack } from '@uifabric/experiments';
+import { IFontStyles, IStyle } from '@uifabric/experiments/lib/Styling';
 import { createStatelessComponent, IStyleableComponentProps, IStatelessComponent } from '@uifabric/experiments/lib/Foundation';
 
 const TestText = 'The quick brown fox jumped over the lazy dog.';
@@ -11,19 +10,8 @@ interface ISetting<TType> {
   usage: string;
 }
 
-const Variants: ISetting<keyof IFontVariants>[] = [
-  { name: 'default', usage: 'default' },
-  { name: 'caption', usage: 'caption' },
-  { name: 'h1', usage: 'h1' },
-  { name: 'h2', usage: 'h2' },
-  { name: 'h3', usage: 'h3' },
-  { name: 'h4', usage: 'h4' },
-  { name: 'h5', usage: 'h5' },
-  { name: 'link', usage: 'link' }
-];
-
-const Sizes: ISetting<keyof IFontSizes>[] = [
-  { name: 'mini', usage: 'usage here.' },
+const Variants: ISetting<keyof IFontStyles>[] = [
+  { name: 'tiny', usage: 'usage here.' },
   { name: 'xSmall', usage: 'usage here.' },
   { name: 'small', usage: 'usage here.' },
   { name: 'smallPlus', usage: 'usage here.' },
@@ -33,24 +21,6 @@ const Sizes: ISetting<keyof IFontSizes>[] = [
   { name: 'xLarge', usage: 'usage here.' },
   { name: 'xxLarge', usage: 'usage here.' },
   { name: 'mega', usage: 'usage here.' }
-];
-
-const Weights: ISetting<keyof IFontWeights>[] = [
-  { name: 'default', usage: '' },
-  { name: 'light', usage: '' },
-  { name: 'regular', usage: '' },
-  { name: 'semibold', usage: '' },
-  { name: 'bold', usage: '' }
-];
-
-const Families: ISetting<keyof IFontFamilies>[] = [{ name: 'default', usage: '' }, { name: 'monospace', usage: '' }];
-
-const Colors: ISetting<keyof ISemanticTextColors | keyof IPalette>[] = [
-  { name: 'bodyText', usage: '' },
-  { name: 'link', usage: '' },
-  { name: 'linkHovered', usage: '' },
-  { name: 'actionLink', usage: '' },
-  { name: 'actionLinkHovered', usage: '' }
 ];
 
 interface ITableStyles {
@@ -70,13 +40,13 @@ interface ITableProps extends IStyleableComponentProps<ITableProps, ITableStyles
 type ITableComponent = IStatelessComponent<ITableProps, ITableStyles>;
 
 const TableView: ITableComponent['view'] = props => (
-  <VerticalStack className={props.className} gap={20}>
-    <Text variant="h3">{props.title}</Text>
+  <Stack className={props.className} gap={20}>
+    <Text variant="medium">{props.title}</Text>
     <table className={props.classNames.table}>
       <thead>
         <tr className={props.classNames.header}>
           {props.headers.map((header: string) => (
-            <Text key={header} as="td" weight="bold">
+            <Text key={header} as="td">
               {header}
             </Text>
           ))}
@@ -84,7 +54,7 @@ const TableView: ITableComponent['view'] = props => (
       </thead>
       <tbody>{props.children}</tbody>
     </table>
-  </VerticalStack>
+  </Stack>
 );
 
 const Table = createStatelessComponent<ITableProps, ITableStyles>({
@@ -115,11 +85,11 @@ const TableRow: React.StatelessComponent<ITableRowProps> = (props: ITableRowProp
 );
 
 export const TextRampExample = () => (
-  <VerticalStack gap={40}>
+  <Stack gap={40}>
     <Text>Default text should render using the "default" variant.</Text>
 
     <Table title="Variants" headers={['Variant', 'Example', 'Usage']}>
-      {Variants.map((setting: ISetting<keyof IFontVariants>) => (
+      {Variants.map((setting: ISetting<keyof IFontStyles>) => (
         <TableRow
           key={setting.name}
           cells={[
@@ -132,64 +102,5 @@ export const TextRampExample = () => (
         />
       ))}
     </Table>
-
-    <Table title="Sizes" headers={['Size', 'Example', 'Usage']}>
-      {Sizes.map((setting: ISetting<keyof IFontSizes>) => (
-        <TableRow
-          key={setting.name}
-          cells={[
-            setting.name,
-            <Text key={setting.name + 'text'} size={setting.name}>
-              {TestText}
-            </Text>,
-            setting.usage
-          ]}
-        />
-      ))}
-    </Table>
-
-    <Table title="Weights" headers={['Weight', 'Example', 'Usage']}>
-      {Weights.map((setting: ISetting<keyof IFontWeights>) => (
-        <TableRow
-          key={setting.name}
-          cells={[
-            setting.name,
-            <Text key={setting.name + 'text'} weight={setting.name}>
-              {TestText}
-            </Text>,
-            setting.usage
-          ]}
-        />
-      ))}
-    </Table>
-
-    <Table title="Families" headers={['Family', 'Example', 'Usage']}>
-      {Families.map((setting: ISetting<keyof IFontFamilies>) => (
-        <TableRow
-          key={setting.name}
-          cells={[
-            setting.name,
-            <Text key={setting.name + 'text'} family={setting.name}>
-              {TestText}
-            </Text>,
-            setting.usage
-          ]}
-        />
-      ))}
-    </Table>
-
-    <Table title="Colors" headers={['Color', 'Example']}>
-      {Colors.map((setting: ISetting<keyof ISemanticTextColors | keyof IPalette>) => (
-        <TableRow
-          key={setting.name}
-          cells={[
-            setting.name,
-            <Text key={setting.name + 'text'} color={setting.name}>
-              {TestText}
-            </Text>
-          ]}
-        />
-      ))}
-    </Table>
-  </VerticalStack>
+  </Stack>
 );