+++ packages/office-ui-fabric-react/src/components/HoverCard/HoverCard.types.ts
@@ -5,7 +5,7 @@ import { IStyle, ITheme } from '../../Styling';
 import { IRefObject, IStyleFunctionOrObject, KeyCodes } from '../../Utilities';
 import { IPlainCardProps } from './PlainCard/PlainCard.types';
 
-export interface IHoverCard { }
+export interface IHoverCard {}
 
 /**
  * HoverCard component props.