+++ packages/office-ui-fabric-react/src/components/pickers/TagPicker/TagItem.tsx
@@ -1,32 +1,39 @@
-/* tslint:disable */
 import * as React from 'react';
-/* tslint:enable */
-import { css, IDictionary } from '../../../Utilities';
+
+import { styled, classNamesFunction } from '../../../Utilities';
 import { Icon } from '../../../Icon';
-import { IPickerItemProps } from '../PickerItem.types';
-import { ITag } from './TagPicker';
-import * as stylesImport from './TagItem.scss';
-const styles: any = stylesImport;
+import { ITagItemProps, ITagItemStyleProps, ITagItemStyles } from './TagPicker.types';
+import { getStyles } from './TagItem.styles';
 
-export interface ITagItemProps extends IPickerItemProps<ITag> {
-  enableTagFocusInDisabledPicker?: boolean;
-}
+const getClassNames = classNamesFunction<ITagItemStyleProps, ITagItemStyles>();
 
-export const TagItem = (props: ITagItemProps) => (
-  <div
-    className={css('ms-TagItem', styles.root, { 'is-selected': props.selected } as IDictionary, props.selected && styles.isSelected)}
-    role={'listitem'}
-    key={props.index}
-    data-selection-index={props.index}
-    data-is-focusable={(props.enableTagFocusInDisabledPicker || !props.disabled) && true}
-  >
-    <span className={css('ms-TagItem-text', styles.tagItemText)} aria-label={props.children as string}>
-      {props.children}
-    </span>
-    {!props.disabled && (
-      <span className={css('ms-TagItem-close', styles.tagItemClose)} onClick={props.onRemoveItem}>
-        <Icon iconName="Cancel" />
+export const TagItemBase = (props: ITagItemProps) => {
+  const { theme, styles, selected, disabled, enableTagFocusInDisabledPicker, children, className, index, onRemoveItem } = props;
+
+  const classNames = getClassNames(styles, {
+    theme: theme!,
+    className,
+    selected
+  });
+
+  return (
+    <div
+      className={classNames.root}
+      role={'listitem'}
+      key={index}
+      data-selection-index={index}
+      data-is-focusable={(enableTagFocusInDisabledPicker || !disabled) && true}
+    >
+      <span className={classNames.text} aria-label={children as string}>
+        {children}
       </span>
-    )}
-  </div>
-);
+      {!disabled && (
+        <span className={classNames.close} onClick={onRemoveItem}>
+          <Icon iconName="Cancel" />
+        </span>
+      )}
+    </div>
+  );
+};
+
+export const TagItem = styled<ITagItemProps, ITagItemStyleProps, ITagItemStyles>(TagItemBase, getStyles, undefined, { scope: 'TagItem' });